# swagger/swagger.yaml
swagger: '2.0'
info:
  description: FlexOffice Demo API
  title: FlexOffice Rest API
  version: "1.0.0"
# host: whispering-lowlands-83497.herokuapp.com
# basePath: /
schemes:
- https
securityDefinitions:
  UserSecurity: 
    type: apiKey
    in: header
    name: authorization
security:
  - UserSecurity: []
paths:
...
  /api/login:
    post:
      consumes:
        - application/json
      produces:
        - application/json; charset=utf-8
      parameters:
        - in: body
          name: body
          schema:
            properties:
              email:
                type: string
              password:
                type: string
            type: object
            example: |-
              {
                  "email": "abc@abc.abc",
                  "password": "abcabc"
              }
      responses:
        '200':
          description: Confirms that the user is connected and returns the token
          schema:
            $ref: '#/definitions/authModel'
      description: Log in
...
  /api/me:
    get:
      produces:
        - application/json; charset=utf-8
      parameters:
        - in: header
          name: authorization
          required: true
          type: string
          x-example: >-
            Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVjYzA1NjU1NTYwYjI0MDAxN2RjNjk5NSIsImlhdCI6MTU1NjEwODg4NX0.SW-2IQeH74mCsE93S27CnRM2B0btjGSJc-PgQICxn34
      responses:
        '200':
          description: Returns the user
          schema:
            $ref: '#/definitions/userModel'
      description: Get user
...
  /api/register:
    post:
      consumes:
        - application/json
      produces:
        - application/json; charset=utf-8
      parameters:
        - in: body
          name: body
          schema:
            properties:
              email:
                type: string
              name:
                type: string
              password:
                type: string
            type: object
      responses:
        '200':
          description: Confirms that the user is connected and returns the token
          schema:
            $ref: '#/definitions/authModel'
      description: Register new user
# TO FIX
...
definitions:
  authModel:
    properties:
      auth:
        type: boolean
      token: {}
    type: object
...
  userModel:
    properties:
      fname:
        type: string
      id:
        type: string
      id_place: {}
      name:
        type: string
      photo: {}
      remoteDay: {}
    type: object
...
host: virtserver.swaggerhub.com
basePath: /Lformery8/flex/1.0.0